:recycle: Split pipeline, train and test files
